authorId 타입변경 -> int

+++ src/main/java/com/nhnacademy/taskapi/author/domain/Author.java
@@ -14,7 +14,7 @@ import org.hibernate.validator.constraints.Length;
 public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long  authorId;
+    private int  authorId;
 
     @NotBlank
     @Length(max=20)
